# src/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Casino en Línea API",
    "description": "API REST del casino en línea. Autenticación via JWT (Bearer token en header `Authorization` o cookie `token`).\n\n## Flujo típico\n1. `POST /api/auth/register` → obtén `token`\n2. Usa `Authorize` (candado arriba) y pega el token\n3. Ya puedes probar el resto de endpoints\n\n## Sockets (Socket.io)\nConecta a `ws://localhost:3000` y emite los siguientes eventos:\n- `joinLobby` → `{ lobbyId, userId, username }` — unirse a la sala de un lobby\n- `leaveLobby` → `{ lobbyId, userId, username }` — salir de la sala\n\nEscucha:\n- `playerJoined` → `{ lobbyId, userId, username }` — otro jugador entró\n- `playerLeft` → `{ lobbyId, userId, username }` — otro jugador salió",
    "version": "1.2.0",
    "contact": {
      "name": "Casino Online"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desarrollo local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenido en login o register. Ejemplo: `Bearer eyJhbGci...`"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "6630a1b2c3d4e5f678901234" },
          "name": { "type": "string", "example": "Manuel" },
          "last_name": { "type": "string", "example": "García" },
          "email": { "type": "string", "format": "email", "example": "manuel@example.com" },
          "phone_number": { "type": "string", "example": "+523311223344" },
          "birthdate": { "type": "string", "format": "date-time" },
          "bank": { "type": "number", "example": 1500 },
          "avatarUrl": { "type": "string", "nullable": true, "example": "https://res.cloudinary.com/..." },
          "authProvider": { "type": "string", "enum": ["local", "google"] },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "Acceso no autorizado. Token requerido." }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Token ausente o inválido",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
      "ServerError": {
        "description": "Error interno del servidor",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Registrar nuevo usuario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name","last_name","email","day","month","year","phone_number","password"],
                "properties": {
                  "name": { "type": "string", "example": "Manuel" },
                  "last_name": { "type": "string", "example": "García" },
                  "email": { "type": "string", "format": "email", "example": "manuel@example.com" },
                  "day": { "type": "integer", "example": 15 },
                  "month": { "type": "integer", "example": 3 },
                  "year": { "type": "integer", "example": 2000 },
                  "phone_number": { "type": "string", "example": "+523311223344" },
                  "password": { "type": "string", "minLength": 6, "example": "mipassword123" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "token": { "type": "string" },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "400": { "description": "Campos faltantes" },
          "409": { "description": "Email ya registrado" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Iniciar sesión",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email","password"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "manuel@example.com" },
                  "password": { "type": "string", "example": "mipassword123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso, retorna JWT",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "token": { "type": "string", "description": "JWT — guárdalo para las demás llamadas" },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "400": { "description": "Campos faltantes" },
          "401": { "description": "Contraseña incorrecta" },
          "404": { "description": "Usuario no encontrado" }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Cerrar sesión (limpia cookie)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Sesión cerrada" }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Obtener usuario autenticado actual",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Datos del usuario",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/auth/google": {
      "get": {
        "tags": ["Auth"],
        "summary": "Iniciar flujo Google OAuth 2.0",
        "description": "Redirige al usuario a Google para autenticar. Requiere `GOOGLE_CLIENT_ID` y `GOOGLE_CLIENT_SECRET` en `.env`. Si no están configurados responde 503.\n\n**No se prueba en Swagger UI** — abrir directamente en el navegador.",
        "responses": {
          "302": { "description": "Redirección a Google" },
          "503": { "description": "OAuth no configurado" }
        }
      }
    },
    "/api/auth/google/callback": {
      "get": {
        "tags": ["Auth"],
        "summary": "Callback Google OAuth (manejado por el servidor)",
        "description": "Google redirige aquí tras autenticar. El servidor emite el JWT en cookie y **redirige al home** (`/`). El token también se incluye en el body para clientes API.",
        "responses": {
          "302": { "description": "Redirect a `/` con cookie `token` seteada" },
          "401": { "description": "Autenticación con Google fallida" },
          "503": { "description": "OAuth no configurado" }
        }
      }
    },
...
  }
}